+++ src/graphql/queries.js
@@ -84,6 +84,7 @@ export const getUser = /* GraphQL */ `
       name
       color
       themePref
+      plexSearchEnabled
       watchlist {
         items {
           userId
@@ -118,6 +119,7 @@ export const listUsers = /* GraphQL */ `
         name
         color
         themePref
+        plexSearchEnabled
         watchlist {
           nextToken
         }
@@ -140,6 +142,7 @@ export const getReview = /* GraphQL */ `
         name
         color
         themePref
+        plexSearchEnabled
         watchlist {
           nextToken
         }
@@ -199,6 +202,7 @@ export const listReviews = /* GraphQL */ `
           name
           color
           themePref
+          plexSearchEnabled
           createdAt
           updatedAt
         }
@@ -300,6 +304,49 @@ export const listShows = /* GraphQL */ `
     }
   }
 `;
+export const watchlistItemsByShow = /* GraphQL */ `
+  query WatchlistItemsByShow(
+    $showId: ID
+    $sortDirection: ModelSortDirection
+    $filter: ModelWatchlistItemFilterInput
+    $limit: Int
+    $nextToken: String
+  ) {
+    watchlistItemsByShow(
+      showId: $showId
+      sortDirection: $sortDirection
+      filter: $filter
+      limit: $limit
+      nextToken: $nextToken
+    ) {
+      items {
+        userId
+        showId
+        show {
+          id
+          tmdbId
+          title
+          type
+          rating
+          img
+          backgroundImg
+          releaseDate
+          description
+          imdbRating
+          rtRating
+          providerIds
+          createdAt
+          source
+          updatedAt
+          owner
+        }
+        createdAt
+        updatedAt
+      }
+      nextToken
+    }
+  }
+`;
 export const reviewsByUser = /* GraphQL */ `
   query ReviewsByUser(
     $userId: ID
@@ -327,6 +374,7 @@ export const reviewsByUser = /* GraphQL */ `
           name
           color
           themePref
+          plexSearchEnabled
           createdAt
           updatedAt
         }