+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED] - 2022-##-##
 
+## [1.19.0] - 2022-05-10
+### Added
+- plex show search button
+- plex search user pref
+
+### Changed
+- svgr config retains viewboxes
+- watchlisted shows no longer removable
+
+### Fixed
+- removing/updating show upon removal of last review
+
 ## [1.18.0] - 2022-05-07
 ### Added
 - trending shows ttl